NIAD-767: Code restructure
