+++ docker/sandbox/app/http_headers.py
@@ -1,0 +1,2 @@
+ANY_CONTENT_TYPE = "*/*"
+FHIR_CONTENT_TYPE = "application/fhir+json"
