+++ docker/sandbox/app/utils.py
@@ -1,0 +1,5 @@
+import uuid
+
+
+def new_uuid():
+    return str(uuid.uuid4()).upper()
