+++ docker/sandbox/main.py
@@ -1,18 +1,16 @@
 import json
 import os
-import re
-import uuid
-from dataclasses import dataclass
 from typing import Optional
-import copy
-from urllib.parse import unquote
 
 from flask import Flask, request, abort, Response
+from app.http_headers import FHIR_CONTENT_TYPE
+from app.accept_header_validator import check_accept_header
+from app.correlation_id_reader import get_correlation_id
+from app.stub_loader import build_empty_bundle, build_bundle, ENDPOINT_RESPONSE_TEMPLATE, ENDPOINT_QUERY_PARAMETERS, \
+    DEVICE_QUERY_PARAMETERS, DEVICE_RESPONSE_TEMPLATE, build_operation_outcome, OPERATION_OUTCOME_400_TEMPLATE, \
+    OPERATION_OUTCOME_404_TEMPLATE, OPERATION_OUTCOME_405_TEMPLATE, OPERATION_OUTCOME_406_TEMPLATE
 
 app = Flask(__name__)
-
-X_CORRELATION_ID = "X-Correlation-ID"
-UUID_PATTERN = "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
 
 ORGANIZATION_QUERY_PARAMETER_NAME = "organization"
 IDENTIFIER_QUERY_PARAMETER_NAME = "identifier"
@@ -22,121 +20,6 @@ ORGANIZATION_FHIR_IDENTIFIER = "https://fhir.nhs.uk/Id/ods-organization-code"
 SERVICE_ID_FHIR_IDENTIFIER = "https://fhir.nhs.uk/Id/nhsEndpointServiceId"
 PARTY_KEY_FHIR_IDENTIFIER = "https://fhir.nhs.uk/Id/nhsMhsPartyKey"
 MANAGING_ORGANIZATION_FHIR_IDENTIFIER = "https://fhir.nhs.uk/Id/ods-organization-code"
-
-FHIR_CONTENT_TYPE = "application/fhir+json"
-
-ENDPOINT_STUB_FILE = "stubs/endpoint.json"
-DEVICE_STUB_FILE = "stubs/device.json"
-
-
-def load_stub(file):
-    with open(file) as f:
-        return json.load(f)
-
-
-@dataclass
-class MatchingDeviceParameters:
-    organization: str
-    service_id: str
-    party_key: str
-    managing_organization: str
-
-    def match(self, organization, service_id, party_key, managing_organization):
-        return organization == self.organization and service_id == self.service_id \
-               and (not party_key and not managing_organization
-                    or managing_organization == self.managing_organization and not party_key
-                    or not managing_organization and party_key == self.party_key
-                    or managing_organization == self.managing_organization and party_key == self.party_key)
-
-
-@dataclass
-class MatchingEndpointParameters:
-    organization: str
-    service_id: str
-    party_key: str
-
-    def match(self, organization, service_id, party_key):
-        return organization == self.organization \
-               and (service_id == self.service_id and not party_key
-                    or not service_id and party_key == self.party_key
-                    or service_id == self.service_id and party_key == self.party_key)
-
-
-_ENDPOINT_STUB = load_stub("stubs/endpoint.json")
-ENDPOINT_QUERY_PARAMETERS = MatchingEndpointParameters(
-    organization=_ENDPOINT_STUB["query_parameters"]["organization"],
-    service_id=_ENDPOINT_STUB["query_parameters"]["service_id"],
-    party_key=_ENDPOINT_STUB["query_parameters"]["party_key"])
-ENDPOINT_RESPONSE_TEMPLATE = _ENDPOINT_STUB["response_body"]
-
-_DEVICE_STUB = load_stub("stubs/device.json")
-DEVICE_QUERY_PARAMETERS = MatchingDeviceParameters(
-    organization=_DEVICE_STUB["query_parameters"]["organization"],
-    service_id=_DEVICE_STUB["query_parameters"]["service_id"],
-    party_key=_DEVICE_STUB["query_parameters"]["party_key"],
-    managing_organization=_DEVICE_STUB["query_parameters"]["managing_organization"])
-DEVICE_RESPONSE_TEMPLATE = _DEVICE_STUB["response_body"]
-
-
-EMPTY_BUNDLE_TEMPLATE = load_stub("stubs/empty_bundle.json")
-OPERATION_OUTCOME_400_TEMPLATE = load_stub("stubs/400.json")
-OPERATION_OUTCOME_404_TEMPLATE = load_stub("stubs/404.json")
-OPERATION_OUTCOME_405_TEMPLATE = load_stub("stubs/405.json")
-OPERATION_OUTCOME_406_TEMPLATE = load_stub("stubs/406.json")
-
-
-def new_uuid():
-    return str(uuid.uuid4()).upper()
-
-
-def build_empty_bundle():
-    bundle = copy.deepcopy(EMPTY_BUNDLE_TEMPLATE)
-    bundle["id"] = new_uuid()
-    bundle["link"][0]["url"] = unquote(request.url)
-    return bundle
-
-
-def build_operation_outcome(operation_outcome_template, diagnostics):
-    operation_outcome = copy.deepcopy(operation_outcome_template)
-    operation_outcome["id"] = new_uuid()
-    if diagnostics:
-        operation_outcome["issue"][0]["diagnostics"] = diagnostics
-    return operation_outcome
-
-
-def build_bundle(bundle_stub_template):
-    bundle = copy.deepcopy(bundle_stub_template)
-    bundle["id"] = new_uuid()
-    bundle["link"][0]["url"] = unquote(request.url)
-
-    for entry in bundle["entry"]:
-        resource_id = new_uuid()
-        entry["resource"]["id"] = resource_id
-        entry["fullUrl"] = f"{request.base_url}/{resource_id}"
-
-    return bundle
-
-
-def get_correlation_id(raise_error=True):
-    correlation_id = request.headers.get(X_CORRELATION_ID, None)
-
-    if not correlation_id:
-        correlation_id = new_uuid()
-    if len(re.findall(UUID_PATTERN, correlation_id)) != 1:
-        if raise_error:
-            abort(400, f"Invalid {X_CORRELATION_ID} header. Should be an UUIDv4 matching regex '{UUID_PATTERN}'")
-        else:
-            correlation_id = new_uuid()
-
-    return correlation_id
-
-
-def check_accept_header():
-    accept_header = request.headers['Accept']
-    if accept_header == "*/*":
-        accept_header = FHIR_CONTENT_TYPE
-    if accept_header != FHIR_CONTENT_TYPE:
-        abort(406)
 
 
 @app.route('/healthcheck')
@@ -211,7 +94,7 @@ def get_optional_query_param(query_param_name: str, fhir_identifier: str) -> Opt
     return result_value
 
 
-def operation_outcome_error_handler(template, status, diagnostics=None):
+def _operation_outcome_error_handler(template, status, diagnostics=None):
     correlation_id = get_correlation_id(raise_error=False)
     operation_outcome = build_operation_outcome(template, diagnostics)
     return Response(
@@ -223,22 +106,22 @@ def operation_outcome_error_handler(template, status, diagnostics=None):
 
 @app.errorhandler(400)
 def bad_request(error):
-    return operation_outcome_error_handler(OPERATION_OUTCOME_400_TEMPLATE, error.code, error.description)
+    return _operation_outcome_error_handler(OPERATION_OUTCOME_400_TEMPLATE, error.code, error.description)
 
 
 @app.errorhandler(404)
 def page_not_found(error):
-    return operation_outcome_error_handler(OPERATION_OUTCOME_404_TEMPLATE, error.code)
+    return _operation_outcome_error_handler(OPERATION_OUTCOME_404_TEMPLATE, error.code)
 
 
 @app.errorhandler(405)
 def method_not_allowed(error):
-    return operation_outcome_error_handler(OPERATION_OUTCOME_405_TEMPLATE, error.code)
+    return _operation_outcome_error_handler(OPERATION_OUTCOME_405_TEMPLATE, error.code)
 
 
 @app.errorhandler(406)
 def not_acceptable(error):
-    return operation_outcome_error_handler(OPERATION_OUTCOME_406_TEMPLATE, error.code)
+    return _operation_outcome_error_handler(OPERATION_OUTCOME_406_TEMPLATE, error.code)
 
 
 if __name__ == '__main__':
